layered::service

Trait Service

pub trait Service<Request> {
    type Response;
    type Error;

    // Required method
    fn call(
        &mut self,
        req: Request,
    ) -> impl Future<Output = Result<Self::Response, Self::Error>> + Send;
}

A trait representing a service that processes requests and produces responses.

Type Parameters

Request - The type of the request being processed.

Required Associated Types

type Response

The type of the response produced by the service.

type Error

The type of the error that may occur during request processing.

Required Methods

fn call( &mut self, req: Request, ) -> impl Future<Output = Result<Self::Response, Self::Error>> + Send

Processes the given request asynchronously.

Arguments

req - The request to be processed.

Returns

A future that resolves to a result containing either the response or an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.